+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
 library(conquestr)
+library(ggplot2)
 
 test_check("conquestr")